+++ src/app/Controllers/ProductController.php
@@ -40,6 +40,11 @@ class ProductController
     }
 
     public function update(): void {
+        if($_SERVER['REQUEST_METHOD'] !== 'POST') {
+            echo "metodo incorreto para a rota";
+            die;
+        }
+
         $id = $_POST['id'];
         $data = [
             'code' => $_POST['code'],
@@ -51,6 +56,11 @@ class ProductController
     }
 
     public function destroy(): void {
+        if($_SERVER['REQUEST_METHOD'] !== 'GET') {
+            echo "metodo incorreto para a rota";
+            die;
+        }
+
         $id = $_GET['delete'];
         $this->model->delete($id);
         $this->reloadPage();
